Add Hidden Group feature
When press "Alt+Tab", the group will not be showed in tabs, if this group
is set to be hidden. This group will be showd only when press the group's
hotkey, and the tabs will show only this group (Use Hide_Other_Group
feature).

Note: Used "EXE" Group for test. Setting file reading logic will be update
in the future.